tests/wpt/web-platform-tests/css/css-animations/parsing/animation-fill-mode-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-fill-mode with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-fill-mode">
<meta name="assert" content="animation-fill-mode supports only the grammar '<single-animation-fill-mode> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("animation-fill-mode", "auto");
test_invalid_value("animation-fill-mode", "forwards backwards");

test_invalid_value("animation-fill-mode", "both, initial");
test_invalid_value("animation-fill-mode", "initial, both");
</script>
</body>
</html>
